requirements.txt
torch>=2.0
torchvision
ultralytics
fastapi
uvicorn[standard]
python-multipart
pydantic>=2.0
langgraph
langchain-core
langchain-openai
rasterio
numpy
Pillow
albumentations
geojson
python-dotenv
httpx
pytest
pytest-asyncio
